+++ src/weather/MainInfo.java
@@ -2,8 +2,8 @@ package weather;
 
 public class MainInfo {
     double temp;
-    int pressure;
-    int humidity;
+    double pressure;
+    double humidity;
     double temp_min;
     double temp_max;
 
@@ -26,19 +26,19 @@ public class MainInfo {
         this.temp = temp;
     }
 
-    public int getPressure() {
+    public double getPressure() {
         return pressure;
     }
 
-    public void setPressure(int pressure) {
+    public void setPressure(double pressure) {
         this.pressure = pressure;
     }
 
-    public int getHumidity() {
+    public double getHumidity() {
         return humidity;
     }
 
-    public void setHumidity(int humidity) {
+    public void setHumidity(double humidity) {
         this.humidity = humidity;
     }
